+++ tests/handlers/test_sliding_sync.py
@@ -18,7 +18,7 @@
 #
 #
 import logging
-from typing import AbstractSet, Dict, List, Optional, Tuple
+from typing import AbstractSet, Dict, Optional, Tuple
 from unittest.mock import patch
 
 from parameterized import parameterized
@@ -26,16 +26,11 @@ from parameterized import parameterized
 from twisted.test.proto_helpers import MemoryReactor
 
 from synapse.api.constants import (
-    AccountDataTypes,
-    EventContentFields,
     EventTypes,
     JoinRules,
     Membership,
-    RoomTypes,
 )
 from synapse.api.room_versions import RoomVersions
-from synapse.events import StrippedStateEvent, make_event_from_dict
-from synapse.events.snapshot import EventContext
 from synapse.handlers.sliding_sync import (
     RoomsForUserType,
     RoomSyncConfig,
@@ -2982,1384 +2977,6 @@ class FilterRoomsRelevantForSyncTestCase(HomeserverTestCase):
         self.assertTrue(room_id1 not in newly_joined)
         # We should be `newly_left` because we were removed via state reset during the from/to range
         self.assertTrue(room_id1 in newly_left)
-
-
-class FilterRoomsTestCase(HomeserverTestCase):
-    """
-    Tests Sliding Sync handler `filter_rooms()` to make sure it includes/excludes rooms
-    correctly.
-    """
-
-    servlets = [
-        admin.register_servlets,
-        knock.register_servlets,
-        login.register_servlets,
-        room.register_servlets,
-    ]
-
-    def default_config(self) -> JsonDict:
-        config = super().default_config()
-        # Enable sliding sync
-        config["experimental_features"] = {"msc3575_enabled": True}
-        return config
-
-    def prepare(self, reactor: MemoryReactor, clock: Clock, hs: HomeServer) -> None:
-        self.sliding_sync_handler = self.hs.get_sliding_sync_handler()
-        self.store = self.hs.get_datastores().main
-        self.event_sources = hs.get_event_sources()
-
-    def _get_sync_room_ids_for_user(
-        self,
-        user: UserID,
-        to_token: StreamToken,
-        from_token: Optional[StreamToken],
-    ) -> Tuple[Dict[str, RoomsForUserType], AbstractSet[str], AbstractSet[str]]:
-        """
-        Get the rooms the user should be syncing with
-        """
-        room_membership_for_user_map, newly_joined, newly_left = self.get_success(
-            self.sliding_sync_handler.room_lists.get_room_membership_for_user_at_to_token(
-                user=user,
-                from_token=from_token,
-                to_token=to_token,
-            )
-        )
-        filtered_sync_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms_relevant_for_sync(
-                user=user,
-                room_membership_for_user_map=room_membership_for_user_map,
-                newly_left_room_ids=newly_left,
-            )
-        )
-
-        return filtered_sync_room_map, newly_joined, newly_left
-
-    def _create_dm_room(
-        self,
-        inviter_user_id: str,
-        inviter_tok: str,
-        invitee_user_id: str,
-        invitee_tok: str,
-    ) -> str:
-        """
-        Helper to create a DM room as the "inviter" and invite the "invitee" user to the room. The
-        "invitee" user also will join the room. The `m.direct` account data will be set
-        for both users.
-        """
-
-        # Create a room and send an invite the other user
-        room_id = self.helper.create_room_as(
-            inviter_user_id,
-            is_public=False,
-            tok=inviter_tok,
-        )
-        self.helper.invite(
-            room_id,
-            src=inviter_user_id,
-            targ=invitee_user_id,
-            tok=inviter_tok,
-            extra_data={"is_direct": True},
-        )
-        # Person that was invited joins the room
-        self.helper.join(room_id, invitee_user_id, tok=invitee_tok)
-
-        # Mimic the client setting the room as a direct message in the global account
-        # data
-        self.get_success(
-            self.store.add_account_data_for_user(
-                invitee_user_id,
-                AccountDataTypes.DIRECT,
-                {inviter_user_id: [room_id]},
-            )
-        )
-        self.get_success(
-            self.store.add_account_data_for_user(
-                inviter_user_id,
-                AccountDataTypes.DIRECT,
-                {invitee_user_id: [room_id]},
-            )
-        )
-
-        return room_id
-
-    _remote_invite_count: int = 0
-
-    def _create_remote_invite_room_for_user(
-        self,
-        invitee_user_id: str,
-        unsigned_invite_room_state: Optional[List[StrippedStateEvent]],
-    ) -> str:
-        """
-        Create a fake invite for a remote room and persist it.
-
-        We don't have any state for these kind of rooms and can only rely on the
-        stripped state included in the unsigned portion of the invite event to identify
-        the room.
-
-        Args:
-            invitee_user_id: The person being invited
-            unsigned_invite_room_state: List of stripped state events to assist the
-                receiver in identifying the room.
-
-        Returns:
-            The room ID of the remote invite room
-        """
-        invite_room_id = f"!test_room{self._remote_invite_count}:remote_server"
-
-        invite_event_dict = {
-            "room_id": invite_room_id,
-            "sender": "@inviter:remote_server",
-            "state_key": invitee_user_id,
-            "depth": 1,
-            "origin_server_ts": 1,
-            "type": EventTypes.Member,
-            "content": {"membership": Membership.INVITE},
-            "auth_events": [],
-            "prev_events": [],
-        }
-        if unsigned_invite_room_state is not None:
-            serialized_stripped_state_events = []
-            for stripped_event in unsigned_invite_room_state:
-                serialized_stripped_state_events.append(
-                    {
-                        "type": stripped_event.type,
-                        "state_key": stripped_event.state_key,
-                        "sender": stripped_event.sender,
-                        "content": stripped_event.content,
-                    }
-                )
-
-            invite_event_dict["unsigned"] = {
-                "invite_room_state": serialized_stripped_state_events
-            }
-
-        invite_event = make_event_from_dict(
-            invite_event_dict,
-            room_version=RoomVersions.V10,
-        )
-        invite_event.internal_metadata.outlier = True
-        invite_event.internal_metadata.out_of_band_membership = True
-
-        self.get_success(
-            self.store.maybe_store_room_on_outlier_membership(
-                room_id=invite_room_id, room_version=invite_event.room_version
-            )
-        )
-        context = EventContext.for_outlier(self.hs.get_storage_controllers())
-        persist_controller = self.hs.get_storage_controllers().persistence
-        assert persist_controller is not None
-        self.get_success(persist_controller.persist_event(invite_event, context))
-
-        self._remote_invite_count += 1
-
-        return invite_room_id
-
-    def test_filter_dm_rooms(self) -> None:
-        """
-        Test `filter.is_dm` for DM rooms
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        # Create a normal room
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        # Create a DM room
-        dm_room_id = self._create_dm_room(
-            inviter_user_id=user1_id,
-            inviter_tok=user1_tok,
-            invitee_user_id=user2_id,
-            invitee_tok=user2_tok,
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        dm_room_ids = self.get_success(
-            self.sliding_sync_handler.room_lists._get_dm_rooms_for_user(user1_id)
-        )
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try with `is_dm=True`
-        truthy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_dm=True,
-                ),
-                after_rooms_token,
-                dm_room_ids=dm_room_ids,
-            )
-        )
-
-        self.assertEqual(truthy_filtered_room_map.keys(), {dm_room_id})
-
-        # Try with `is_dm=False`
-        falsy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_dm=False,
-                ),
-                after_rooms_token,
-                dm_room_ids=dm_room_ids,
-            )
-        )
-
-        self.assertEqual(falsy_filtered_room_map.keys(), {room_id})
-
-    def test_filter_encrypted_rooms(self) -> None:
-        """
-        Test `filter.is_encrypted` for encrypted rooms
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create an unencrypted room
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        # Create an encrypted room
-        encrypted_room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-        self.helper.send_state(
-            encrypted_room_id,
-            EventTypes.RoomEncryption,
-            {EventContentFields.ENCRYPTION_ALGORITHM: "m.megolm.v1.aes-sha2"},
-            tok=user1_tok,
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try with `is_encrypted=True`
-        truthy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=True,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(truthy_filtered_room_map.keys(), {encrypted_room_id})
-
-        # Try with `is_encrypted=False`
-        falsy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=False,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(falsy_filtered_room_map.keys(), {room_id})
-
-    def test_filter_encrypted_server_left_room(self) -> None:
-        """
-        Test that we can apply a `filter.is_encrypted` against a room that everyone has left.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        before_rooms_token = self.event_sources.get_current_token()
-
-        # Create an unencrypted room
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-        # Leave the room
-        self.helper.leave(room_id, user1_id, tok=user1_tok)
-
-        # Create an encrypted room
-        encrypted_room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-        self.helper.send_state(
-            encrypted_room_id,
-            EventTypes.RoomEncryption,
-            {EventContentFields.ENCRYPTION_ALGORITHM: "m.megolm.v1.aes-sha2"},
-            tok=user1_tok,
-        )
-        # Leave the room
-        self.helper.leave(encrypted_room_id, user1_id, tok=user1_tok)
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            # We're using a `from_token` so that the room is considered `newly_left` and
-            # appears in our list of relevant sync rooms
-            from_token=before_rooms_token,
-            to_token=after_rooms_token,
-        )
-
-        # Try with `is_encrypted=True`
-        truthy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=True,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(truthy_filtered_room_map.keys(), {encrypted_room_id})
-
-        # Try with `is_encrypted=False`
-        falsy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=False,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(falsy_filtered_room_map.keys(), {room_id})
-
-    def test_filter_encrypted_server_left_room2(self) -> None:
-        """
-        Test that we can apply a `filter.is_encrypted` against a room that everyone has
-        left.
-
-        There is still someone local who is invited to the rooms but that doesn't affect
-        whether the server is participating in the room (users need to be joined).
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        _user2_tok = self.login(user2_id, "pass")
-
-        before_rooms_token = self.event_sources.get_current_token()
-
-        # Create an unencrypted room
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-        # Invite user2
-        self.helper.invite(room_id, targ=user2_id, tok=user1_tok)
-        # User1 leaves the room
-        self.helper.leave(room_id, user1_id, tok=user1_tok)
-
-        # Create an encrypted room
-        encrypted_room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-        self.helper.send_state(
-            encrypted_room_id,
-            EventTypes.RoomEncryption,
-            {EventContentFields.ENCRYPTION_ALGORITHM: "m.megolm.v1.aes-sha2"},
-            tok=user1_tok,
-        )
-        # Invite user2
-        self.helper.invite(encrypted_room_id, targ=user2_id, tok=user1_tok)
-        # User1 leaves the room
-        self.helper.leave(encrypted_room_id, user1_id, tok=user1_tok)
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            # We're using a `from_token` so that the room is considered `newly_left` and
-            # appears in our list of relevant sync rooms
-            from_token=before_rooms_token,
-            to_token=after_rooms_token,
-        )
-
-        # Try with `is_encrypted=True`
-        truthy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=True,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(truthy_filtered_room_map.keys(), {encrypted_room_id})
-
-        # Try with `is_encrypted=False`
-        falsy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=False,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(falsy_filtered_room_map.keys(), {room_id})
-
-    def test_filter_encrypted_after_we_left(self) -> None:
-        """
-        Test that we can apply a `filter.is_encrypted` against a room that was encrypted
-        after we left the room (make sure we don't just use the current state)
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        before_rooms_token = self.event_sources.get_current_token()
-
-        # Create an unencrypted room
-        room_id = self.helper.create_room_as(user2_id, tok=user2_tok)
-        # Leave the room
-        self.helper.join(room_id, user1_id, tok=user1_tok)
-        self.helper.leave(room_id, user1_id, tok=user1_tok)
-
-        # Create a room that will be encrypted
-        encrypted_after_we_left_room_id = self.helper.create_room_as(
-            user2_id, tok=user2_tok
-        )
-        # Leave the room
-        self.helper.join(encrypted_after_we_left_room_id, user1_id, tok=user1_tok)
-        self.helper.leave(encrypted_after_we_left_room_id, user1_id, tok=user1_tok)
-
-        # Encrypt the room after we've left
-        self.helper.send_state(
-            encrypted_after_we_left_room_id,
-            EventTypes.RoomEncryption,
-            {EventContentFields.ENCRYPTION_ALGORITHM: "m.megolm.v1.aes-sha2"},
-            tok=user2_tok,
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            # We're using a `from_token` so that the room is considered `newly_left` and
-            # appears in our list of relevant sync rooms
-            from_token=before_rooms_token,
-            to_token=after_rooms_token,
-        )
-
-        # Try with `is_encrypted=True`
-        truthy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=True,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # Even though we left the room before it was encrypted, we still see it because
-        # someone else on our server is still participating in the room and we "leak"
-        # the current state to the left user. But we consider the room encryption status
-        # to not be a secret given it's often set at the start of the room and it's one
-        # of the stripped state events that is normally handed out.
-        self.assertEqual(
-            truthy_filtered_room_map.keys(), {encrypted_after_we_left_room_id}
-        )
-
-        # Try with `is_encrypted=False`
-        falsy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=False,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # Even though we left the room before it was encrypted... (see comment above)
-        self.assertEqual(falsy_filtered_room_map.keys(), {room_id})
-
-    def test_filter_encrypted_with_remote_invite_room_no_stripped_state(self) -> None:
-        """
-        Test that we can apply a `filter.is_encrypted` filter against a remote invite
-        room without any `unsigned.invite_room_state` (stripped state).
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create a remote invite room without any `unsigned.invite_room_state`
-        _remote_invite_room_id = self._create_remote_invite_room_for_user(
-            user1_id, None
-        )
-
-        # Create an unencrypted room
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        # Create an encrypted room
-        encrypted_room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-        self.helper.send_state(
-            encrypted_room_id,
-            EventTypes.RoomEncryption,
-            {EventContentFields.ENCRYPTION_ALGORITHM: "m.megolm.v1.aes-sha2"},
-            tok=user1_tok,
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try with `is_encrypted=True`
-        truthy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=True,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should not appear because we can't figure out whether
-        # it is encrypted or not (no stripped state, `unsigned.invite_room_state`).
-        self.assertEqual(truthy_filtered_room_map.keys(), {encrypted_room_id})
-
-        # Try with `is_encrypted=False`
-        falsy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=False,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should not appear because we can't figure out whether
-        # it is encrypted or not (no stripped state, `unsigned.invite_room_state`).
-        self.assertEqual(falsy_filtered_room_map.keys(), {room_id})
-
-    def test_filter_encrypted_with_remote_invite_encrypted_room(self) -> None:
-        """
-        Test that we can apply a `filter.is_encrypted` filter against a remote invite
-        encrypted room with some `unsigned.invite_room_state` (stripped state).
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create a remote invite room with some `unsigned.invite_room_state`
-        # indicating that the room is encrypted.
-        remote_invite_room_id = self._create_remote_invite_room_for_user(
-            user1_id,
-            [
-                StrippedStateEvent(
-                    type=EventTypes.Create,
-                    state_key="",
-                    sender="@inviter:remote_server",
-                    content={
-                        EventContentFields.ROOM_CREATOR: "@inviter:remote_server",
-                        EventContentFields.ROOM_VERSION: RoomVersions.V10.identifier,
-                    },
-                ),
-                StrippedStateEvent(
-                    type=EventTypes.RoomEncryption,
-                    state_key="",
-                    sender="@inviter:remote_server",
-                    content={
-                        EventContentFields.ENCRYPTION_ALGORITHM: "m.megolm.v1.aes-sha2",
-                    },
-                ),
-            ],
-        )
-
-        # Create an unencrypted room
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        # Create an encrypted room
-        encrypted_room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-        self.helper.send_state(
-            encrypted_room_id,
-            EventTypes.RoomEncryption,
-            {EventContentFields.ENCRYPTION_ALGORITHM: "m.megolm.v1.aes-sha2"},
-            tok=user1_tok,
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try with `is_encrypted=True`
-        truthy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=True,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should appear here because it is encrypted
-        # according to the stripped state
-        self.assertEqual(
-            truthy_filtered_room_map.keys(), {encrypted_room_id, remote_invite_room_id}
-        )
-
-        # Try with `is_encrypted=False`
-        falsy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=False,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should not appear here because it is encrypted
-        # according to the stripped state
-        self.assertEqual(falsy_filtered_room_map.keys(), {room_id})
-
-    def test_filter_encrypted_with_remote_invite_unencrypted_room(self) -> None:
-        """
-        Test that we can apply a `filter.is_encrypted` filter against a remote invite
-        unencrypted room with some `unsigned.invite_room_state` (stripped state).
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create a remote invite room with some `unsigned.invite_room_state`
-        # but don't set any room encryption event.
-        remote_invite_room_id = self._create_remote_invite_room_for_user(
-            user1_id,
-            [
-                StrippedStateEvent(
-                    type=EventTypes.Create,
-                    state_key="",
-                    sender="@inviter:remote_server",
-                    content={
-                        EventContentFields.ROOM_CREATOR: "@inviter:remote_server",
-                        EventContentFields.ROOM_VERSION: RoomVersions.V10.identifier,
-                    },
-                ),
-                # No room encryption event
-            ],
-        )
-
-        # Create an unencrypted room
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        # Create an encrypted room
-        encrypted_room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-        self.helper.send_state(
-            encrypted_room_id,
-            EventTypes.RoomEncryption,
-            {EventContentFields.ENCRYPTION_ALGORITHM: "m.megolm.v1.aes-sha2"},
-            tok=user1_tok,
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try with `is_encrypted=True`
-        truthy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=True,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should not appear here because it is unencrypted
-        # according to the stripped state
-        self.assertEqual(truthy_filtered_room_map.keys(), {encrypted_room_id})
-
-        # Try with `is_encrypted=False`
-        falsy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_encrypted=False,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should appear because it is unencrypted according to
-        # the stripped state
-        self.assertEqual(
-            falsy_filtered_room_map.keys(), {room_id, remote_invite_room_id}
-        )
-
-    def test_filter_invite_rooms(self) -> None:
-        """
-        Test `filter.is_invite` for rooms that the user has been invited to
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        # Create a normal room
-        room_id = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.join(room_id, user1_id, tok=user1_tok)
-
-        # Create a room that user1 is invited to
-        invite_room_id = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.invite(invite_room_id, src=user2_id, targ=user1_id, tok=user2_tok)
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try with `is_invite=True`
-        truthy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_invite=True,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(truthy_filtered_room_map.keys(), {invite_room_id})
-
-        # Try with `is_invite=False`
-        falsy_filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    is_invite=False,
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(falsy_filtered_room_map.keys(), {room_id})
-
-    def test_filter_room_types(self) -> None:
-        """
-        Test `filter.room_types` for different room types
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create a normal room (no room type)
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        # Create a space room
-        space_room_id = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-            extra_content={
-                "creation_content": {EventContentFields.ROOM_TYPE: RoomTypes.SPACE}
-            },
-        )
-
-        # Create an arbitrarily typed room
-        foo_room_id = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-            extra_content={
-                "creation_content": {
-                    EventContentFields.ROOM_TYPE: "org.matrix.foobarbaz"
-                }
-            },
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try finding only normal rooms
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[None]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {room_id})
-
-        # Try finding only spaces
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[RoomTypes.SPACE]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {space_room_id})
-
-        # Try finding normal rooms and spaces
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    room_types=[None, RoomTypes.SPACE]
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {room_id, space_room_id})
-
-        # Try finding an arbitrary room type
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    room_types=["org.matrix.foobarbaz"]
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {foo_room_id})
-
-    def test_filter_not_room_types(self) -> None:
-        """
-        Test `filter.not_room_types` for different room types
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create a normal room (no room type)
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        # Create a space room
-        space_room_id = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-            extra_content={
-                "creation_content": {EventContentFields.ROOM_TYPE: RoomTypes.SPACE}
-            },
-        )
-
-        # Create an arbitrarily typed room
-        foo_room_id = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-            extra_content={
-                "creation_content": {
-                    EventContentFields.ROOM_TYPE: "org.matrix.foobarbaz"
-                }
-            },
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try finding *NOT* normal rooms
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(not_room_types=[None]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {space_room_id, foo_room_id})
-
-        # Try finding *NOT* spaces
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    not_room_types=[RoomTypes.SPACE]
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {room_id, foo_room_id})
-
-        # Try finding *NOT* normal rooms or spaces
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    not_room_types=[None, RoomTypes.SPACE]
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {foo_room_id})
-
-        # Test how it behaves when we have both `room_types` and `not_room_types`.
-        # `not_room_types` should win.
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    room_types=[None], not_room_types=[None]
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # Nothing matches because nothing is both a normal room and not a normal room
-        self.assertEqual(filtered_room_map.keys(), set())
-
-        # Test how it behaves when we have both `room_types` and `not_room_types`.
-        # `not_room_types` should win.
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(
-                    room_types=[None, RoomTypes.SPACE], not_room_types=[None]
-                ),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {space_room_id})
-
-    def test_filter_room_types_server_left_room(self) -> None:
-        """
-        Test that we can apply a `filter.room_types` against a room that everyone has left.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        before_rooms_token = self.event_sources.get_current_token()
-
-        # Create a normal room (no room type)
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-        # Leave the room
-        self.helper.leave(room_id, user1_id, tok=user1_tok)
-
-        # Create a space room
-        space_room_id = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-            extra_content={
-                "creation_content": {EventContentFields.ROOM_TYPE: RoomTypes.SPACE}
-            },
-        )
-        # Leave the room
-        self.helper.leave(space_room_id, user1_id, tok=user1_tok)
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            # We're using a `from_token` so that the room is considered `newly_left` and
-            # appears in our list of relevant sync rooms
-            from_token=before_rooms_token,
-            to_token=after_rooms_token,
-        )
-
-        # Try finding only normal rooms
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[None]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {room_id})
-
-        # Try finding only spaces
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[RoomTypes.SPACE]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {space_room_id})
-
-    def test_filter_room_types_server_left_room2(self) -> None:
-        """
-        Test that we can apply a `filter.room_types` against a room that everyone has left.
-
-        There is still someone local who is invited to the rooms but that doesn't affect
-        whether the server is participating in the room (users need to be joined).
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        _user2_tok = self.login(user2_id, "pass")
-
-        before_rooms_token = self.event_sources.get_current_token()
-
-        # Create a normal room (no room type)
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-        # Invite user2
-        self.helper.invite(room_id, targ=user2_id, tok=user1_tok)
-        # User1 leaves the room
-        self.helper.leave(room_id, user1_id, tok=user1_tok)
-
-        # Create a space room
-        space_room_id = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-            extra_content={
-                "creation_content": {EventContentFields.ROOM_TYPE: RoomTypes.SPACE}
-            },
-        )
-        # Invite user2
-        self.helper.invite(space_room_id, targ=user2_id, tok=user1_tok)
-        # User1 leaves the room
-        self.helper.leave(space_room_id, user1_id, tok=user1_tok)
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            # We're using a `from_token` so that the room is considered `newly_left` and
-            # appears in our list of relevant sync rooms
-            from_token=before_rooms_token,
-            to_token=after_rooms_token,
-        )
-
-        # Try finding only normal rooms
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[None]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {room_id})
-
-        # Try finding only spaces
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[RoomTypes.SPACE]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        self.assertEqual(filtered_room_map.keys(), {space_room_id})
-
-    def test_filter_room_types_with_remote_invite_room_no_stripped_state(self) -> None:
-        """
-        Test that we can apply a `filter.room_types` filter against a remote invite
-        room without any `unsigned.invite_room_state` (stripped state).
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create a remote invite room without any `unsigned.invite_room_state`
-        _remote_invite_room_id = self._create_remote_invite_room_for_user(
-            user1_id, None
-        )
-
-        # Create a normal room (no room type)
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        # Create a space room
-        space_room_id = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-            extra_content={
-                "creation_content": {EventContentFields.ROOM_TYPE: RoomTypes.SPACE}
-            },
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try finding only normal rooms
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[None]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should not appear because we can't figure out what
-        # room type it is (no stripped state, `unsigned.invite_room_state`)
-        self.assertEqual(filtered_room_map.keys(), {room_id})
-
-        # Try finding only spaces
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[RoomTypes.SPACE]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should not appear because we can't figure out what
-        # room type it is (no stripped state, `unsigned.invite_room_state`)
-        self.assertEqual(filtered_room_map.keys(), {space_room_id})
-
-    def test_filter_room_types_with_remote_invite_space(self) -> None:
-        """
-        Test that we can apply a `filter.room_types` filter against a remote invite
-        to a space room with some `unsigned.invite_room_state` (stripped state).
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create a remote invite room with some `unsigned.invite_room_state` indicating
-        # that it is a space room
-        remote_invite_room_id = self._create_remote_invite_room_for_user(
-            user1_id,
-            [
-                StrippedStateEvent(
-                    type=EventTypes.Create,
-                    state_key="",
-                    sender="@inviter:remote_server",
-                    content={
-                        EventContentFields.ROOM_CREATOR: "@inviter:remote_server",
-                        EventContentFields.ROOM_VERSION: RoomVersions.V10.identifier,
-                        # Specify that it is a space room
-                        EventContentFields.ROOM_TYPE: RoomTypes.SPACE,
-                    },
-                ),
-            ],
-        )
-
-        # Create a normal room (no room type)
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        # Create a space room
-        space_room_id = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-            extra_content={
-                "creation_content": {EventContentFields.ROOM_TYPE: RoomTypes.SPACE}
-            },
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try finding only normal rooms
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[None]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should not appear here because it is a space room
-        # according to the stripped state
-        self.assertEqual(filtered_room_map.keys(), {room_id})
-
-        # Try finding only spaces
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[RoomTypes.SPACE]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should appear here because it is a space room
-        # according to the stripped state
-        self.assertEqual(
-            filtered_room_map.keys(), {space_room_id, remote_invite_room_id}
-        )
-
-    def test_filter_room_types_with_remote_invite_normal_room(self) -> None:
-        """
-        Test that we can apply a `filter.room_types` filter against a remote invite
-        to a normal room with some `unsigned.invite_room_state` (stripped state).
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create a remote invite room with some `unsigned.invite_room_state`
-        # but the create event does not specify a room type (normal room)
-        remote_invite_room_id = self._create_remote_invite_room_for_user(
-            user1_id,
-            [
-                StrippedStateEvent(
-                    type=EventTypes.Create,
-                    state_key="",
-                    sender="@inviter:remote_server",
-                    content={
-                        EventContentFields.ROOM_CREATOR: "@inviter:remote_server",
-                        EventContentFields.ROOM_VERSION: RoomVersions.V10.identifier,
-                        # No room type means this is a normal room
-                    },
-                ),
-            ],
-        )
-
-        # Create a normal room (no room type)
-        room_id = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        # Create a space room
-        space_room_id = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-            extra_content={
-                "creation_content": {EventContentFields.ROOM_TYPE: RoomTypes.SPACE}
-            },
-        )
-
-        after_rooms_token = self.event_sources.get_current_token()
-
-        # Get the rooms the user should be syncing with
-        sync_room_map, newly_joined, newly_left = self._get_sync_room_ids_for_user(
-            UserID.from_string(user1_id),
-            from_token=None,
-            to_token=after_rooms_token,
-        )
-
-        # Try finding only normal rooms
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[None]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should appear here because it is a normal room
-        # according to the stripped state (no room type)
-        self.assertEqual(filtered_room_map.keys(), {room_id, remote_invite_room_id})
-
-        # Try finding only spaces
-        filtered_room_map = self.get_success(
-            self.sliding_sync_handler.room_lists.filter_rooms(
-                UserID.from_string(user1_id),
-                sync_room_map,
-                SlidingSyncConfig.SlidingSyncList.Filters(room_types=[RoomTypes.SPACE]),
-                after_rooms_token,
-                dm_room_ids=set(),
-            )
-        )
-
-        # `remote_invite_room_id` should not appear here because it is a normal room
-        # according to the stripped state (no room type)
-        self.assertEqual(filtered_room_map.keys(), {space_room_id})
 
 
 class SortRoomsTestCase(HomeserverTestCase):
